+++ package/src/components/AudioPlayer/Interface/Information/TrackInfo.tsx
@@ -1,26 +1,41 @@
-import { useTrackContext } from "@/hooks/context/useTrackContext";
+import { useTrackContext } from "@/components/AudioPlayer/Context/hooks/useTrackContext";
 import { FC, memo } from "react";
+import Grid, { GridItemLayoutProps } from "@/components/Grid";
+import { useResolvedGridArea } from "../hooks/useResolvedGridArea";
 import "./TrackInfo.css";
 
-export const TrackInfo: FC = memo(function TrackInfo() {
+export type TrackInfoProps = GridItemLayoutProps;
+
+export const TrackInfo: FC<TrackInfoProps> = memo(function TrackInfo({
+  gridArea,
+  visible,
+  ...rest
+}) {
   const { playList, curIdx } = useTrackContext();
   const curPlayData = playList[curIdx];
+
+  const resolvedGridArea = useResolvedGridArea("trackInfo", gridArea);
+
   return (
-    <div className="rmap-track-info-container">
-      {curPlayData?.customTrackInfo ? (
-        curPlayData.customTrackInfo
-      ) : (
-        <>
-          {curPlayData?.name && (
-            <span className="title" data-testid="track-title">
-              {curPlayData.name}
-            </span>
-          )}
-          {curPlayData?.writer && (
-            <span className="writer">{curPlayData.writer}</span>
-          )}
-        </>
-      )}
-    </div>
+    <Grid.Item gridArea={resolvedGridArea} visible={visible ?? true} {...rest}>
+      <div className="rmap-track-info-container">
+        {curPlayData?.customTrackInfo ? (
+          curPlayData.customTrackInfo
+        ) : (
+          <>
+            {curPlayData?.name && (
+              <span className="title" data-testid="track-title">
+                {curPlayData.name}
+              </span>
+            )}
+            {curPlayData?.writer && (
+              <span className="writer">{curPlayData.writer}</span>
+            )}
+          </>
+        )}
+      </div>
+    </Grid.Item>
   );
 });
+
+TrackInfo.displayName = "TrackInfo";