+++ package/src/components/AudioPlayer/Interface/Information/index.tsx
@@ -1,7 +1,5 @@
-import Grid from "@/components/Grid";
-import { defaultInterfacePlacement } from "@/components/AudioPlayer/Context/StateContext";
-import { useTrackContext } from "@/hooks/context/useTrackContext";
-import { useUIContext } from "@/hooks/context/useUIContext";
+import { useTrackContext } from "@/components/AudioPlayer/Context/hooks/useTrackContext";
+import { useUIContext } from "@/components/AudioPlayer/Context/hooks/useUIContext";
 import { FC } from "react";
 import { Artwork } from "./Artwork";
 import { TrackInfo } from "./TrackInfo";
@@ -9,7 +7,7 @@ import { TrackTime } from "./TrackTime";
 
 export const Information: FC = () => {
   const { playList, curIdx } = useTrackContext();
-  const { interfacePlacement, activeUI } = useUIContext();
+  const { activeUI } = useUIContext();
 
   const isTrackInfoActive =
     Boolean(
@@ -20,30 +18,13 @@ export const Information: FC = () => {
 
   return (
     <>
-      <Grid.Item
-        gridArea={
-          interfacePlacement?.itemCustomArea?.artwork ||
-          interfacePlacement?.templateArea?.artwork ||
-          defaultInterfacePlacement.templateArea.artwork
-        }
+      <Artwork
         visible={Boolean(
           playList[curIdx]?.img && (activeUI.artwork ?? activeUI.all)
         )}
-      >
-        <Artwork />
-      </Grid.Item>
-      <Grid.Item
-        gridArea={
-          interfacePlacement?.itemCustomArea?.trackInfo ||
-          interfacePlacement?.templateArea?.trackInfo ||
-          defaultInterfacePlacement.templateArea.trackInfo
-        }
-        visible={isTrackInfoActive}
-      >
-        <TrackInfo />
-      </Grid.Item>
-
-      <TrackTime />
+      />
+      <TrackInfo visible={isTrackInfoActive} />
+      <TrackTime visible={Boolean(activeUI.trackTime ?? activeUI.all)} />
     </>
   );
 };